docs/history/v0.2.1/src/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vault-mcp"
version = "0.2.1"
description = "MCP-compliant Obsidian documentation server with RAG and live sync"
authors = [{name = "Vault MCP Team", email = "team@vault-mcp.dev"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "watchdog>=3.0.0",
    "mistune>=3.0.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "toml>=0.10.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.0.0",
    "litellm>=1.28.0",
    "openai>=1.12.0",
    "mlx-lm>=0.12.0",
    "llama-index-core>=0.10.0",
    "llama-index-llms-litellm",
    "llama-index-vector-stores-chroma",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "bandit>=1.7.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
vault-mcp = "components.mcp_server.main:main"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
target-version = "py311"
line-length = 88
src = ["components", "vault_mcp", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "B", "SIM", "I"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ARG002", "F841", "SIM117", "B018"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
exclude = ["tests/"]
namespace_packages = true
explicit_package_bases = true

[tool.bandit]
exclude_dirs = ["tests", "components/file_watcher/tests", "components/mcp_server/tests", "components/vector_store/tests"]
skips = ["B101"]

[tool.pytest.ini_options]
testpaths = ["tests", "components/*/tests"]
python_files = ["test_*.py"]
addopts = "--cov=vault_mcp --cov=components --cov-report=term-missing"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "types-toml>=0.10.8.20240310",
]
